+++ frontend/src/pages/SignIn.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { signin, parseJwt } from "../api";
+import { signin, parseJwt } from "@/services/authApi"
 import { useNavigate } from "react-router-dom";
 
 export default function SignIn() {
